% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idWorkspace.R
\name{idWorkspace}
\alias{idWorkspace}
\title{List and select available Objective Connect workspaces}
\usage{
idWorkspace()
}
\description{
List and select available Objective Connect workspaces
}
\examples{
idWorkspace()

}
\keyword{objective}
\keyword{objr}
