% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushConnect.R
\name{pushConnect}
\alias{pushConnect}
\title{Move an existing file into the Objective Connect workspace}
\usage{
pushConnect(doc, name, update = F)
}
\arguments{
\item{doc}{path to file to be sent to the Objective Connect workspace}

\item{name}{name of the new object in the Objective Connect workspace. Not required for updating existing objects.}

\item{update}{update and exists document? (logical)}
}
\description{
Move an existing file into the Objective Connect workspace
}
\examples{
pushConnect("file.xlsx","New File Name")

}
\keyword{objective}
\keyword{objr}
